Move class definitions for `Domain` and `Euclidean` into an internal module
This will prevent dependency loops when I add new classes between
`Domain` and `Euclidean` and give them constrained-default
implementations based on `Euclidean`.
